pyproject.toml
requires-python = ">=3.12"
[project]
name = "olist_analytics"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas",
    "polars",
    "sqlalchemy",
    "matplotlib",
    "seaborn",
    "jupyterlab",
    "python-dotenv",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-bigquery-storage>=2.35.0",
    "dbt-core>=1.10.15",
    "dbt-bigquery>=1.10.3",
    "openai>=2.11.0",
]
